Move RequestConfiguration, RequestConfigurer to internal package.
