+++ src/test/java/coresearch/cvurl/io/model/ConfigurationTest.java
@@ -1,6 +1,7 @@
 package coresearch.cvurl.io.model;
 
 import coresearch.cvurl.io.constant.HttpClientMode;
+import coresearch.cvurl.io.internal.configuration.RequestConfiguration;
 import coresearch.cvurl.io.mapper.GenericMapper;
 import coresearch.cvurl.io.mapper.impl.JacksonMapper;
 import org.junit.jupiter.api.Test;
